[Feat] #83 - Move error from unoffical SDK.
